Fix chapter title alignment, refactor
