+++ patches/src/main/kotlin/app/morphe/patches/youtube/interaction/seekbar/SeekbarThumbnailPreviewPatch.kt
@@ -10,34 +10,22 @@ package app.morphe.patches.youtube.interaction.seekbar
 import app.morphe.patcher.extensions.InstructionExtensions.addInstruction
 import app.morphe.patcher.extensions.InstructionExtensions.addInstructions
 import app.morphe.patcher.extensions.InstructionExtensions.addInstructionsWithLabels
-import app.morphe.patcher.extensions.InstructionExtensions.getInstruction
 import app.morphe.patcher.patch.bytecodePatch
-import app.morphe.patcher.util.proxy.mutableTypes.MutableMethod.Companion.toMutable
 import app.morphe.patches.all.misc.resources.resourceMappingPatch
 import app.morphe.patches.shared.misc.settings.preference.SwitchPreference
+import app.morphe.patches.youtube.misc.chapters.chaptersHookPatch
 import app.morphe.patches.youtube.misc.extension.sharedExtensionPatch
 import app.morphe.patches.youtube.misc.playservice.is_21_12_or_greater
 import app.morphe.patches.youtube.misc.playservice.versionCheckPatch
 import app.morphe.patches.youtube.misc.settings.PreferenceScreen
 import app.morphe.patches.youtube.misc.settings.settingsPatch
 import app.morphe.patches.youtube.shared.Constants.COMPATIBILITY_YOUTUBE
-import app.morphe.patches.youtube.video.videoid.hookVideoId
-import app.morphe.patches.youtube.video.videoid.videoIdPatch
-import app.morphe.util.findFieldFromToString
 import app.morphe.util.getReference
-import app.morphe.util.indexOfFirstInstructionReversedOrThrow
-import com.android.tools.smali.dexlib2.AccessFlags
-import com.android.tools.smali.dexlib2.Opcode
-import com.android.tools.smali.dexlib2.builder.MutableMethodImplementation
-import com.android.tools.smali.dexlib2.iface.instruction.OneRegisterInstruction
 import com.android.tools.smali.dexlib2.iface.instruction.ReferenceInstruction
 import com.android.tools.smali.dexlib2.iface.reference.FieldReference
 import com.android.tools.smali.dexlib2.iface.reference.MethodReference
-import com.android.tools.smali.dexlib2.immutable.ImmutableMethod
 
 private const val EXTENSION_CLASS = "Lapp/morphe/extension/youtube/patches/SeekbarThumbnailPreviewPatch;"
-internal const val EXTENSION_TIMELINE_MARKER_INTERFACE =
-    $$"Lapp/morphe/extension/youtube/patches/SeekbarThumbnailPreviewPatch$TimelineMarker;"
 
 val seekbarThumbnailPreviewPatch = bytecodePatch(
     description = "Adds an option to restore the seekbar thumbnail preview."
@@ -47,7 +35,7 @@ val seekbarThumbnailPreviewPatch = bytecodePatch(
         settingsPatch,
         versionCheckPatch,
         resourceMappingPatch,
-        videoIdPatch
+        chaptersHookPatch
     )
 
     compatibleWith(COMPATIBILITY_YOUTUBE)
@@ -113,91 +101,5 @@ val seekbarThumbnailPreviewPatch = bytecodePatch(
                 nop
             """
         )
-
-        hookVideoId("$EXTENSION_CLASS->newVideoLoaded(Ljava/lang/String;)V")
-
-        getTimelineMarkersArrayFingerprint(
-            TimelineMarkerFingerprint.classDef.type
-        ).method.apply {
-            val index = indexOfFirstInstructionReversedOrThrow(Opcode.RETURN_OBJECT)
-            val register = getInstruction<OneRegisterInstruction>(index).registerA
-            addInstruction(
-                index,
-                "invoke-static { v$register }, $EXTENSION_CLASS->" +
-                        "setTimelineMarkers([$EXTENSION_TIMELINE_MARKER_INTERFACE)V"
-            )
-        }
-
-        TimelineMarkerFingerprint.let {
-            it.classDef.apply {
-                interfaces.add(EXTENSION_TIMELINE_MARKER_INTERFACE)
-
-                val startMillis = it.method.findFieldFromToString("startMillis=")
-                methods.add(
-                    ImmutableMethod(
-                        type,
-                        "patch_getStartMillis",
-                        listOf(),
-                        "J",
-                        AccessFlags.PUBLIC.value or AccessFlags.FINAL.value,
-                        null,
-                        null,
-                        MutableMethodImplementation(2),
-                    ).toMutable().apply {
-                        addInstructionsWithLabels(
-                            0,
-                            """
-                                iget-wide v0, p0, $startMillis
-                                return-wide v0
-                            """
-                        )
-                    }
-                )
-
-                val endMillis = it.method.findFieldFromToString("endMillis=")
-                methods.add(
-                    ImmutableMethod(
-                        type,
-                        "patch_getEndMillis",
-                        listOf(),
-                        "J",
-                        AccessFlags.PUBLIC.value or AccessFlags.FINAL.value,
-                        null,
-                        null,
-                        MutableMethodImplementation(2),
-                    ).toMutable().apply {
-                        addInstructionsWithLabels(
-                            0,
-                            """
-                                iget-wide v0, p0, $endMillis
-                                return-wide v0
-                            """
-                        )
-                    }
-                )
-
-                val title = it.method.findFieldFromToString("title=")
-                methods.add(
-                    ImmutableMethod(
-                        type,
-                        "patch_getTitle",
-                        listOf(),
-                        "Ljava/lang/CharSequence;",
-                        AccessFlags.PUBLIC.value or AccessFlags.FINAL.value,
-                        null,
-                        null,
-                        MutableMethodImplementation(2),
-                    ).toMutable().apply {
-                        addInstructionsWithLabels(
-                            0,
-                            """
-                                iget-object v0, p0, $title
-                                return-object v0
-                            """
-                        )
-                    }
-                )
-            }
-        }
     }
 }
